+++ gemclus/gemini/_fdivergences.py
@@ -182,7 +182,7 @@ class HellingerGEMINI(_FDivergence):
     r"""
     Implements the one-vs-all and one-vs-one Squared Hellinger distance GEMINI.
 
-    The one-vs-all version compares the squared Hellinger distance distance between a cluster distribution
+    The one-vs-all version compares the squared Hellinger distance between a cluster distribution
     and the data distribution.
 
     .. math::
@@ -243,11 +243,22 @@ class ChiSquareGEMINI(_FDivergence):
     r"""
     Implements the one-vs-all and one-vs-one Chi Squared divergence GEMINI.
 
+    The one-vs-all version compares the chi square divergence between a cluster distribution
+    and the data distribution.
+
     .. math::
         \mathcal{I} = \mathbb{E}_{y \sim p(y)}[D_{\chi^2}(p(x|y)\|p(x))]
 
+    The one-vs-one version compares the chi square divergence between two cluster distributions.
+
+    .. math::
+        \mathcal{I} = \mathbb{E}_{y_a,y_b \sim p(y)}[D_{\chi^2}(p(x|y_a)\|p(x|y_b))]
+
     Parameters
     ----------
+    ovo: bool, default=False
+        Whether to use the one-vs-all objective (False) or the one-vs-one objective (True).
+
     epsilon: float, default=1e-12
         The precision for clipping the prediction values in order to avoid numerical instabilities.
 
@@ -261,11 +272,13 @@ class ChiSquareGEMINI(_FDivergence):
 
     @constraint_params(
         {
+            "ovo": [bool],
             "epsilon": [Interval(Real, 0, 1, closed="neither")]
         }
     )
-    def __init__(self, epsilon=1e-12):
+    def __init__(self, ovo=False, epsilon=1e-12):
         super().__init__(epsilon)
+        self.ovo = ovo
 
     def evaluate(self, y_pred, affinity, return_grad=False):
         # Use a clip mask for numerical stability in gradients
@@ -274,11 +287,24 @@ class ChiSquareGEMINI(_FDivergence):
         p_y = p_y_x.mean(0)
 
         cluster_wise_estimates = p_y_x / p_y
-        chi2_gemini = np.sum(p_y_x*cluster_wise_estimates, axis=1).mean()
+        if self.ovo:
+            alpha = np.sum(p_y_x*cluster_wise_estimates, axis=1, keepdims=True)
+            beta = np.sum(p_y/cluster_wise_estimates, axis=1, keepdims=True)
+            chi2_gemini = np.mean(alpha*beta)
+        else:
+            chi2_gemini = np.sum(p_y_x*cluster_wise_estimates, axis=1).mean()
 
         if return_grad:
-            gradients = (2*cluster_wise_estimates-np.square(cluster_wise_estimates).mean(0))
+            if self.ovo:
+                single_beta = beta * cluster_wise_estimates
+                double_beta = single_beta * cluster_wise_estimates
+                single_alpha = alpha / cluster_wise_estimates
+                double_alpha = single_alpha / cluster_wise_estimates
+
+                gradients = 2*single_beta-double_alpha  + np.mean(2*single_alpha-double_beta, axis=0)
+            else:
+                gradients = (2*cluster_wise_estimates-np.square(cluster_wise_estimates).mean(0))
             gradients /= y_pred.shape[0]
-            return chi2_gemini, gradients * clip_mask
+            return 0.5*chi2_gemini, 0.5*gradients * clip_mask
         else:
-            return chi2_gemini
+            return 0.5*chi2_gemini